SymbolicExpression: accept key-val pairs as args
