+++ src/core/utils/benchmark_tests/functions/4C_utils_symbolic_expression_benchmark.cpp
@@ -20,11 +20,10 @@ namespace
   {
     // A baseline benchmark which has no variables and only evaluates a constant expression.
     SymbolicExpression<double> expr("3.14");
-    std::map<std::string, double> variables;
 
     for (auto _ : state)
     {
-      double result = expr.value(variables);
+      double result = expr.value();
       benchmark::DoNotOptimize(result);
     }
   }
@@ -35,10 +34,9 @@ namespace
     // A benchmark that evaluates a expression which can be simplified to a constant.
     // Should run as fast as the constant benchmark above.
     SymbolicExpression<double> expr("2*1.0^2 + 2.0 + 4*3.0");
-    std::map<std::string, double> variables;
     for (auto _ : state)
     {
-      double result = expr.value(variables);
+      double result = expr.value();
       benchmark::DoNotOptimize(result);
     }
   }
@@ -48,11 +46,10 @@ namespace
   {
     // Only basic arithmetic, including expensive division.
     SymbolicExpression<double> expr("2*x + y/2 - 4*z");
-    std::map<std::string, double> variables = {{"x", 1.0}, {"y", 2.0}, {"z", 3.0}};
 
     for (auto _ : state)
     {
-      double result = expr.value(variables);
+      double result = expr.value("x", 1.0, "y", 2.0, "z", 3.0);
       benchmark::DoNotOptimize(result);
     }
   }
@@ -79,15 +76,11 @@ namespace
   void symbolic_expression_first_derivative(benchmark::State& state)
   {
     SymbolicExpression<double> expr("2*x^2 + y + 4*z");
-    std::map<std::string, FirstDerivativeType> variables = {
-        {"x", FirstDerivativeType(2, 0, 1.0)},
-        {"y", FirstDerivativeType(2, 1, 2.0)},
-    };
-    std::map<std::string, double> constants = {{"z", 3.0}};
 
     for (auto _ : state)
     {
-      auto result = expr.first_derivative(variables, constants);
+      auto result = expr.first_derivative(
+          "x", FirstDerivativeType(2, 0, 1.0), "y", FirstDerivativeType(2, 1, 2.0), "z", 3.0);
       benchmark::DoNotOptimize(result);
     }
   }
@@ -97,15 +90,11 @@ namespace
   void symbolic_expression_second_derivative(benchmark::State& state)
   {
     SymbolicExpression<double> expr("2*x^2 + y + 4*z");
-    std::map<std::string, SecondDerivativeType> variables = {
-        {"x", SecondDerivativeType(2, 0, 1.0)},
-        {"y", SecondDerivativeType(2, 1, 2.0)},
-    };
-    std::map<std::string, double> constants = {{"z", 3.0}};
 
     for (auto _ : state)
     {
-      auto result = expr.second_derivative(variables, constants);
+      auto result = expr.second_derivative(
+          "x", SecondDerivativeType(2, 0, 1.0), "y", SecondDerivativeType(2, 1, 2.0), "z", 3.0);
       benchmark::DoNotOptimize(result);
     }
   }
@@ -115,11 +104,10 @@ namespace
   void symbolic_expression_functions(benchmark::State& state)
   {
     SymbolicExpression<double> expr("sin(x) + cos(y) * exp(z) - log10(x + y)");
-    std::map<std::string, double> variables = {{"x", 1.0}, {"y", 2.0}, {"z", 3.0}};
 
     for (auto _ : state)
     {
-      double result = expr.value(variables);
+      double result = expr.value("x", 1.0, "y", 2.0, "z", 3.0);
       benchmark::DoNotOptimize(result);
     }
   }
